fix 등록 / 조회 / 수정 / 삭제
fix 등록 / 조회 / 수정 / 삭제

+++ franchise-project-settings/src/main/java/com/ktds/fr/prdt/service/PrdtServiceImpl.java
@@ -1,6 +1,9 @@
 package com.ktds.fr.prdt.service;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.List;
+import java.util.UUID;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,6 +12,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.ktds.fr.common.api.exceptions.ApiArgsException;
 import com.ktds.fr.common.api.exceptions.ApiException;
 import com.ktds.fr.prdt.dao.PrdtDAO;
 import com.ktds.fr.prdt.vo.PrdtVO;
@@ -37,22 +41,34 @@ public class PrdtServiceImpl implements PrdtService {
 	@Override
 	public boolean create(PrdtVO prdtVO, MultipartFile uploadFile) {
 		
-		// TODO 파일 업로드 대기...완성되면 주석해제할것 
-//		if (uploadFile != null && !uploadFile.isEmpty()) {
-//			File dir = new File(profilePath);
-//			if (!dir.exists()) {
-//				dir.mkdirs();
-//			}
-//			String uuidFileName = UUID.randomUUID().toString();
-//			File profileFile = new File(dir, uuidFileName);
-//			try {
-//				uploadFile.transferTo(profileFile);
-//			} catch (IllegalStateException | IOException e) {
-//				logger.error(e.getMessage(), e);
-//			}
-//			prdtVO.setPrdtFileId(uuidFileName);
-//		}
-		prdtVO.setPrdtFileId("sample_file_name");
+		String srt = prdtVO.getPrdtSrt();
+		if (srt == null || srt.trim().length() == 0) {
+			throw new ApiArgsException("400", "분류 선택 필요");
+		}
+		String nm = prdtVO.getPrdtNm();
+		if (nm == null || nm.trim().length() == 0) {
+			throw new ApiArgsException("400", "이름이 비었음");
+		}
+		int prc = prdtVO.getPrdtPrc();
+		if (prc== 0) {
+			throw new ApiArgsException("400", "가격이 비었음");
+		}
+		
+		
+		if (uploadFile != null && !uploadFile.isEmpty()) {
+			File dir = new File(profilePath);
+			if (!dir.exists()) {
+				dir.mkdirs();
+			}
+			String uuidFileName = UUID.randomUUID().toString();
+			File profileFile = new File(dir, uuidFileName);
+			try {
+				uploadFile.transferTo(profileFile);
+			} catch (IllegalStateException | IOException e) {
+				logger.error(e.getMessage(), e);
+			}
+			prdtVO.setPrdtFileId(uuidFileName);
+		}
 		
 		return prdtDAO.create(prdtVO) > 0;
 	}
@@ -60,43 +76,65 @@ public class PrdtServiceImpl implements PrdtService {
 	@Override
 	public boolean update(PrdtVO prdtVO, MultipartFile uploadFile) {
 		
+		String srt = prdtVO.getPrdtSrt();
+		if (srt == null || srt.trim().length() == 0) {
+			throw new ApiArgsException("400", "분류 선택 필요");
+		}
+		String nm = prdtVO.getPrdtNm();
+		if (nm == null || nm.trim().length() == 0) {
+			throw new ApiArgsException("400", "이름이 비었음");
+		}
+		int prc = prdtVO.getPrdtPrc();
+		if (prc== 0) {
+			throw new ApiArgsException("400", "가격이 비었음");
+		}
+		
+		
 		boolean isModify = false;
 		PrdtVO origin= prdtDAO.readOne(prdtVO.getPrdtId());
-		if (!prdtVO.getPrdtNm().equals(origin.getPrdtNm())) {
+		if (!origin.getPrdtNm().equals(prdtVO.getPrdtNm())) {
 			isModify = true;
 		}
-		if (prdtVO.getPrdtPrc() != origin.getPrdtPrc()) {
+		if (origin.getPrdtPrc() != prdtVO.getPrdtPrc()) {
 			isModify = true;
 		}
-		if (!prdtVO.getPrdtNm().equals(origin.getPrdtNm())) {
+		if (!origin.getPrdtNm().equals(prdtVO.getPrdtNm())) {
 			isModify = true;
 		}
-		if (!prdtVO.getPrdtSrt().equals(origin.getPrdtSrt())) {
+		if (!origin.getPrdtSrt().equals(prdtVO.getPrdtSrt())) {
 			isModify = true;
 		}
-		if (!prdtVO.getUseYn().equals(origin.getUseYn())) {
+		String useYn = prdtVO.getUseYn() == null ? "N" : prdtVO.getUseYn() ;
+		if (!origin.getUseYn().equals(useYn)) {
 			isModify = true;
 		}
 		
-		// TODO 파일 업로드 대기...완성되면 주석해제할것 
-//		if (uploadFile != null && !uploadFile.isEmpty()) {
-//			File dir = new File(profilePath);
-//			if (!dir.exists()) {
-//				dir.mkdirs();
-//			}
-//			String uuidFileName = UUID.randomUUID().toString();
-//			File profileFile = new File(dir, uuidFileName);
-//			try {
-//				uploadFile.transferTo(profileFile);
-//			} catch (IllegalStateException | IOException e) {
-//				logger.error(e.getMessage(), e);
-//			}
-//			prdtVO.setPrdtFileId(uuidFileName);
-//		}
-		
-		prdtVO.setPrdtFileId("sample_file_name");
+		if ((prdtVO.getPrdtFileId() == null 
+				|| prdtVO.getPrdtFileId().length() == 0)
+				&& prdtVO.getIsDeleteIMG().equals("N")) {
+			prdtVO.setPrdtFileId(origin.getPrdtFileId());
+		} else {
+			isModify = true;
+		}
+
 		
 		if (isModify) {
+			if (uploadFile != null && !uploadFile.isEmpty()) {
+				File dir = new File(profilePath);
+				if (!dir.exists()) {
+					dir.mkdirs();
+				}
+				String uuidFileName = UUID.randomUUID().toString();
+				File profileFile = new File(dir, uuidFileName);
+				try {
+					uploadFile.transferTo(profileFile);
+				} catch (IllegalStateException | IOException e) {
+					logger.error(e.getMessage(), e);
+				}
+				prdtVO.setPrdtFileId(uuidFileName);
+			} else {
+				prdtVO.setPrdtFileId("none");
+			}
 			return prdtDAO.update(prdtVO) > 0;
 		} else {
 			throw new ApiException("400", "변경된 정보 없음");
@@ -106,11 +144,18 @@ public class PrdtServiceImpl implements PrdtService {
 
 	@Override
 	public boolean deleteOne(String prdtId) {
+		if (prdtId == null || prdtId.trim().length() == 0) {
+			throw new ApiArgsException("400", "선택된 항목이 없습니다.");
+		}
 		return prdtDAO.deleteOne(prdtId) > 0;
 	}
 
 	@Override
 	public boolean deleteSelectAll(List<String> prdtIdList) {
+		System.out.println(prdtIdList.size() + "일괄삭제용 리스트크기다");
+		if (prdtIdList.size()==0) {
+			throw new ApiArgsException("400", "선택된 항목이 없습니다.");
+		}
 		return prdtDAO.deleteSelectAll(prdtIdList) > 0;
 	}
